Final commit of work on sure scheme inc. gutter icons
This was as far as I could get implementing gutter icons even. Even working on this in my own time I couldn't get it to work. From here I will be removing gutter icon and java parser related code to leave a clean verson with the work I did manage to implement successfully.